refactor code to check more than talend - see full commit message
Thanks to more configuration offered, this project can now be used to
monitor other task than talend ones.

Thus the project will change it's name to reflect the new beaviour.

This commit introduce some new configuration option to run the script,
as well as the new name for the package.